+++ pyviews/core/__init__.py
@@ -1,7 +1,7 @@
 """Package for creation instances from xml and its values to expressions"""
 
 from .error import ViewInfo, CoreError
-from .compilation import Expression, ObjectNode, CompilationError, expression
+from .compilation import Expression, ObjectNode, CompilationError
 from .binding import Binding, BindingRule, BindingTarget, BindingError
 from .node import Node, InstanceNode, Modifier, Property
 from .observable import Observable, ObservableEntity, InheritedDict