+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Dev
+
+- Expression resolved using code as resolve parameter
+- Migrated to injectool 1.1.1
+
 ## 2.1.1
 
 - changed call modifier